lexiloop-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-multipart==0.0.6
pydantic==2.5.3